+++ R/bayesfactor_parameters.R
@@ -252,7 +252,8 @@ bayesfactor_parameters.brmsfit <- bayesfactor_parameters.stanreg
 
 
 #' @importFrom stats update
-#' @importFrom insight get_parameters
+#' @importFrom emmeans ref_grid
+#' @importFrom emmeans as.mcmc.emmGrid
 #' @rdname bayesfactor_parameters
 #' @export
 bayesfactor_parameters.emmGrid <- function(posterior, prior = NULL,
@@ -271,8 +272,9 @@ bayesfactor_parameters.emmGrid <- function(posterior, prior = NULL,
     )
   } else if (!inherits(prior, "emmGrid")) { # then is it a model
     prior <- .update_to_priors(prior, verbose = verbose)
-    prior <- insight::get_parameters(prior, effects = "fixed")
-    prior <- stats::update(posterior, post.beta = as.matrix(prior))
+    prior <- emmeans::ref_grid(prior)
+    prior <- prior@post.beta
+    prior <- stats::update(posterior, post.beta = prior)
   }
 
   prior <- as.data.frame(as.matrix(emmeans::as.mcmc.emmGrid(prior, names = FALSE)))